modify Items Structure and ItemInit class
