+++ Game/Items/Item/HealBall.java
@@ -3,7 +3,6 @@ package Game.Items.Item;
 import Game.Items.ItemBall;
 
 import java.awt.Color;
-import java.awt.Image;
 
 public class HealBall extends ItemBall {
 
@@ -11,12 +10,6 @@ public class HealBall extends ItemBall {
 	public HealBall(int x, int y, int width, int height) {
 		super(x, y, width, height);
 		this.colorItem = Color.red;
-	}
-
-
-	/** Constructeur pour la liste d'initialisation des items */
-	public HealBall(int nbMaxItem, int percentItem, Color colorItem, Image imageItem) {
-		super(nbMaxItem, percentItem, colorItem, imageItem);
 	}
