+++ Game/Items/Item/SpeedProjectileBall.java
@@ -3,7 +3,6 @@ package Game.Items.Item;
 import Game.Items.ItemBall;
 
 import java.awt.Color;
-import java.awt.Image;
 
 public class SpeedProjectileBall extends ItemBall {
 
@@ -11,12 +10,6 @@ public class SpeedProjectileBall extends ItemBall {
 	public SpeedProjectileBall(int x, int y, int width, int height) {
 		super(x, y, width, height);
 		this.colorItem = Color.magenta;
-	}
-
-
-	/** Constructeur pour la liste d'initialisation des items */
-	public SpeedProjectileBall(int nbMaxItem, int percentItem, Color colorItem, Image imageItem) {
-		super(nbMaxItem, percentItem, colorItem, imageItem);
 	}
